// services/discord/enums.ts
export enum DiscordApplicationCommandNames {
  GoodMorning = 'goodmorning',
}

export enum DiscordResponses {
  ForecastNotFound = 'The forecast could not be found.',
  GoodMorning = 'Good morning!',
  GoodMorningTeam = 'Good morning team!',
  IDontKnow = "I don't know what to do.",
  LocationNotFound = 'The location could not be found.',
  MissingOptions = 'The command is missing options.',
}